dom/alarm/test/test_alarm_permitted_app.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test Permitted Application for Alarm API</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
<script type="application/javascript">

"use strict";

SpecialPowers.addPermission("alarms", true, document);

// mozAlarms is installed on all platforms except Android for the moment.
if (navigator.appVersion.indexOf("Android") != -1) {
  ok('AlarmsManager' in window, "Interface AlarmsManager should exist");
  try {
    todo('mozAlarms' in navigator,
         "mozAlarms is not allowed on Android for now. TODO Bug 863557.");
  } catch (e) {
    todo(!e, "('mozAlarms' in navigator) should not throw exceptions once " +
             "mozAlarms is installed on Android. TODO Bug 863557. " +
             "Caught exception: " + e);
  }
} else {
  ok('mozAlarms' in navigator, "navigator.mozAlarms should exist");
  ok(navigator.mozAlarms instanceof AlarmsManager,
     "navigator.mozAlarms should be an instance of AlarmsManager");
}

</script>
</pre>
</body>
</html>
